requirements.txt
qdrant-client>=1.10.0
torch
transformers
pillow
requests